[arccore+arcane] Move 'Atomic.h' and files relatives to views from 'arcane_accelerator' to 'arccore_accelerator'.
